Refactor main.ts: Extract large functions into focused modules
This refactoring splits main.ts into smaller, more maintainable modules
following the single responsibility principle.

**Changes:**

**Phase 5: Drag-and-Drop Manager** (~100 lines)
- Created `src/lib/drag-drop-manager.ts` to encapsulate all drag-drop logic
- Moved drag state variables and event handlers from main.ts
- Provides `setupDragAndDrop()` function for easy integration

**Phase 6: Workspace Lifecycle** (~302 lines)
- Created `src/lib/workspace-lifecycle.ts` for workspace initialization
- Extracted `handleWorkspacePathInput()` with dependency injection
- Handles workspace validation, config loading, terminal creation, and agent launching

**Phase 7: Terminal Lifecycle** (~275 lines)
- Created `src/lib/terminal-lifecycle.ts` for terminal management
- Extracted `launchAgentsForTerminals()` for agent launching
- Extracted `reconnectTerminals()` for session recovery
- Extracted `verifyTerminalSessions()` for health checks

**Benefits:**
- Reduced main.ts from 1998 to ~1300 lines
- Better separation of concerns
- Easier to test individual modules with dependency injection
- Improved code organization and maintainability

**Testing:**
- All existing tests pass
- pnpm check:ci completes successfully
- No functional changes, pure refactoring

Closes #283
